Reset %path-procedure-result _after_ evaluating page handlers
%path-procedure-result was being reset in render-page, and that caused
problems when threads were reused and render-page was not called
(e.g., redirect to login page when using sessions).

%path-procedure-result can't be reset in reset-per-request-params
because it is first set at path matching time (before run-resource is
called), so its value would be reset before page handlers could use it
(in render-page). So, we now reset it _after_ the page handler is
finished.