Make awful-start properly honor the dev-mode? keyword parameter
Before this change, development-mode-actions would only be called from
load-apps.  So, if an application uses awful-start with dev-mode? set
to #t, but doesn't call load-apps, development-mode-actions would not
be called.

While at it, call development-mode-actions _before_ evaluating the
thunk for awful-start, so that the evaluated code can override the
settings made by development-mode-actions.